criacao do componente passenger details
